ui/recipes-chat-chainlit/requirements.txt
python-dotenv==1.0.0
pandas==1.5.3   
matplotlib==3.8.2
geopandas==0.14.1
requests==2.31.0
psycopg2_binary==2.9.9
geopandas==0.10.2   
geoalchemy2==0.14.6
